Rewrote tokenizers to follow new conventions
